Fixed visibility of abstract class constructor in HASCOBuilder. Renamed ListBasedRequiredInterfaceSolutionProductionTester.
